+++ CrawlerProject-final/CrawlerApplication/it_conference/scheduler.py
@@ -15,6 +15,6 @@ class Scraper:
 
     def run_spiders(self):
         self.process.crawl(self.spider)
-        self.scheduler.add_job(self.process.crawl, 'interval', args=[self.spider], seconds=300)
+        self.scheduler.add_job(self.process.crawl, 'interval', args=[self.spider], seconds=60)
         self.scheduler.start()
         self.process.start(False)
