+++ CrawlerProject-final/CrawlerProject_API/CrawlerProject_API.csproj
@@ -16,7 +16,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <Folder Include="Models\" />
+    <Folder Include="Migrations\" />
   </ItemGroup>
 
 </Project>
