+++ CrawlerProject-final/CrawlerProject_API/Models/LocalUser.cs
@@ -1,0 +1,11 @@
+﻿namespace CrawlerProject_API.Models
+{
+    public class LocalUser
+    {
+        public int Id { get; set; }
+        public string? UserName { get; set; }
+        public string? Name { get; set; }
+        public string? Password { get; set; }
+        public string? Role { get; set; }
+    }
+}
