+++ CrawlerProject-final/CrawlerProject_API/Repository/IRepository/IConferenceRepository.cs
@@ -1,0 +1,10 @@
+﻿using CrawlerProject_API.Models;
+
+namespace CrawlerProject_API.Repository.IRepository
+{
+    public interface IConferenceRepository
+    {
+        Task<List<Conference>> GetAll();
+        Task Save();
+    }
+}
