+++ CrawlerProject-final/CrawlerProject_API/appsettings.Development.json
@@ -7,6 +7,6 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "CrawlerProjectDb": "Server=tcp:hotelprojectserver.database.windows.net,1433;Initial Catalog=HotelAPI;Persist Security Info=False;User ID=sqladmin;Password=Asddsaas1#;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Connection Timeout=30;"
+    "CrawlerProjectDb": "Server=localhost,1433;Initial Catalog=conferences;Persist Security Info=False;User ID=sa;Password=Password123;MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=True;"
   }
 }
